+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v1.5.0 - 2019-10-31
+
+- Added Copper difficulty as a stepping stone between Bronze and Silver
+- Added Easier Start option which adds more neutral systems with free tech
+
 ## v1.4.1 - 2019-10-24
 
 - Added support for the tutorial personality tag in custom difficulty mode